// apps/api/src/models/types/public/Monitors.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { Id } from './Users';
import type { default as MonitorType } from './MonitorType';
import type { default as MonitorStatus } from './MonitorStatus';

/** Identifier type for monitors */
export type Id = string;

/** Represents the table public.monitors */
export default interface Monitors {
  id: Id;

  author: Id;

  monitor_type: MonitorType;

  name: string;

  description: string | null;

  created_at: Date;

  updated_at: Date;

  schedule: string;

  status: MonitorStatus;

  config: unknown | null;
}

/** Represents the initializer for the table public.monitors */
export interface MonitorsInitializer {
  id: Id;

  author: Id;

  monitor_type: MonitorType;

  name: string;

  description?: string | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  updated_at?: Date;

  schedule: string;

  /** Default value: 'active'::monitor_status */
  status?: MonitorStatus;

  config?: unknown | null;
}

/** Represents the mutator for the table public.monitors */
export interface MonitorsMutator {
  id?: Id;

  author?: Id;

  monitor_type?: MonitorType;

  name?: string;

  description?: string | null;

  created_at?: Date;

  updated_at?: Date;

  schedule?: string;

  status?: MonitorStatus;

  config?: unknown | null;
}